requirements.txt
amqp==5.0.7
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.8
django-admin-geomap==1.1
django-admin-tools==0.9.1
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-filter==21.1
django-leaflet==0.28.2
django-leaflet-admin-list==0.0.4
django-redis==5.1.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-writable-nested==0.6.3
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2==2.9.2
pycodestyle==2.8.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
